requirements.txt
binpacking==1.5.2
colorama==0.4.6
hurry.filesize==0.9
motor==2.5.1
prometheus-client==0.16.0
protobuf >= 3.13.0, < 4.0
pycycle==0.0.8
pyjwt[crypto]==2.5.0
pymongo==3.12.3
requests==2.28.2
tornado==6.2
wipac-dev-tools==1.6.12
wipac-rest-tools==1.4.14
wipac-telemetry==0.2.5
